LayoutTests/fast/events/touch/ios/touchlist-iterator-expected.txt

Tests that TouchList objects have an iterator

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Symbol.iterator in touchList is true
PASS touch is touch1
PASS 'entries' in touchList is false
PASS 'keys' in touchList is false
PASS 'forEach' in touchList is false
PASS 'values' in touchList is false
PASS successfullyParsed is true

TEST COMPLETE
